refactor(upk-objects): split static_mesh along types vs parse seams
Split the 1076-line static_mesh.rs (over the 700-line hard cap in
CLAUDE.md "File organization") into a static_mesh/ module:

- static_mesh/mod.rs        — module docs + re-exports (public surface)
- static_mesh/types.rs      — decoded structs (StaticMesh, BoundingBox,
                              LodModel, Vertex, MeshSection, KdopTriangle)
                              + the collision_triangles accessor + its tests
- static_mesh/parse/mod.rs  — byte-level deserializer + size constants
- static_mesh/parse/tests.rs — parser unit tests

Behavior-preserving · re-export discipline · zero public-surface change.

Part of #529.
